requirements.txt
amqp
appnope
asgiref
asttokens
backcall
billiard.0
cachetools
celery[redis]
certifi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
decorator
Deprecated
distlib
Django
django-extensions
djangorestframework
executing
filelock
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
httplib2
idna
jedi
kombu
matplotlib-inline
mysqlclient
oauthlib
packaging
pexpect
pickleshare
platformdirs
prompt-toolkit
protobuf
ptyprocess
pure-eval
pyasn1
pyasn1-modules
Pygments
pyparsing
pytz
requests
requests-oauthlib
rsa
six
sqlparse
stack-data
traitlets
uritemplate
urllib3
vine
virtualenv
wcwidth
wrapt
